Reorganize packages and extract classes
