+++ src/main/java/duke/command/DukeCommand.java
@@ -1,4 +1,4 @@
-package command;
+package duke.command;
 
 
 public enum DukeCommand {
