+++ src/main/java/duke/tasks/Event.java
@@ -1,7 +1,7 @@
-package tasks;
+package duke.tasks;
 
 import java.time.LocalDate;
-import date.DukeDate;
+import duke.date.DukeDate;
 
 public class Event extends Task {
     private static final long serialVersionUID = 7664438936982546960L;
