+++ src/main/java/duke/ui/Ui.java
@@ -1,8 +1,6 @@
-import driver.Driver;
-import storage.Storage;
-import tasks.TaskList;
+package duke.ui;
 
-public class Duke {
+public class Ui {
 
     private static final String BAR =
             "    ____________________________________________________________";
@@ -10,7 +8,7 @@ public class Duke {
     private static final String INDENTATION = "     ";
     private static final String NEW_LINE = "\n";
 
-    private static void greet() {
+    public static void greet() {
         System.out.println(BAR);
         System.out.println(INDENTATION + "Hello! I'm Duke");
         System.out.println(INDENTATION + "What can I do for you?");
@@ -18,11 +16,5 @@ public class Duke {
         System.out.print(NEW_LINE);
     }
 
-    public static void main(String[] args) {
-        greet();
-        // Storage storage = new Storage();
-        TaskList taskList = Storage.readTaskList();
-        Driver driver = new Driver();
-        driver.run(taskList);
-    }
+
 }
